+++ dapps/shop/backend/utils/events.js
@@ -1,4 +1,3 @@
-// const { PROVIDER } = require('./const')
 const Web3 = require('web3')
 const w3 = new Web3()
